Kind of hacky way to close the connection immediately when client closes
This was suggested on the twisted ticket that was submitted. Seems to work
but, tests do not pass and not sure what the performance hit is on this.

I noticed that on an empty queue when a request comes in while no data is
present in the queue and the client closes right away. The connection persists.